Add option to disable preloading of videos (#63)
* Add option to disable preload of all items

In some scenarios the bandwidth is limited and preloading all the items in the menu takes a lot of time, for example when using a cellular connection. This might slow down the UI or seeing the preview of all videos is just not needed (as in my case).
With an option 'video_preload: false' a static icon can be displayed instead and speed up the page loading. This is not a breaking change as nothing changes if the option is not present in the config.

* Update README.md